raise file descriptor limits for esync
This increases limits of number of opened files by a process. The main
benefactor is Proton/Wine esync which uses a lot of file descriptors in
order to work properly. The limit value is raised from 4K to 256K, which
is in line with a future systemd change (256K will be the default since
systemd 240). See the links in the configuration file for more
information.